% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_C.R
\encoding{UTF-8}
\name{calc_C_p}
\alias{calc_C_p}
\title{Calculate Student Challenge as the Percent Likelihood of Non-Success}
\usage{
calc_C_p(x, S, prev.ss, demos = NULL, clear = T)
}
\arguments{
\item{x}{A data.frame containing student-level data.}

\item{S}{A vector of \eqn{n_{student}} 0/1 student success indicators.}

\item{prev.ss}{A character vector of previous-year scaled score values variable names.}

\item{demos}{An optional character vector of demographic variable names.}

\item{clear}{A logical scalar indicating whether to clear the console for user updates.}

\item{year}{Character scalar name of the variable in \code{x} containing years.}

\item{subject}{Character scalar name of the variable in \code{x} containing subjects.}

\item{grade}{Character scalar name of the variable in \code{x} containing grades.}
}
\value{
A numeric vector of \eqn{n_{student}} values in the interval \eqn{\left[0, 100\right]}.
}
\description{
Calculate Student Challenge as the Percent Likelihood of Non-Success
}
