% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_Ss_from_ALs.R
\encoding{UTF-8}
\name{calc_Cs_P1}
\alias{calc_Cs_P1}
\title{Calculate Probability-Based Challenge Measures from Current-Year Success Indicators and Previous-Year Scaled Scores}
\usage{
calc_Cs_P1(x, S, prev.ss, square = T, clear = T)
}
\arguments{
\item{x}{A data frame containing at least one current-year success indicator variable and at least one previous-year scaled score variable.}

\item{S}{A character vector of the names of variables in \code{x} containing current-year success indicators.}

\item{prev.ss}{A character vector of the names of variables in \code{x} containing previous-year scaled scores.}

\item{square}{A logical scalar indicating whether to square the multiple variables in \code{prev.ss} in the logistic regression predicting each success indicator variable.}

\item{clear}{A logical scalar indicating whether to clear the console for major user updates.}
}
\value{
A data frame with as many new student challenge variables as there are elements in the argument \code{Ss}.
}
\description{
Calculate Probability-Based Challenge Measures from Current-Year Success Indicators and Previous-Year Scaled Scores
}
