% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_descriptives.R
\encoding{UTF-8}
\name{calc_descriptives}
\alias{calc_descriptives}
\title{Attach Input Descriptive Statistics to Output}
\usage{
calc_descriptives(x, Cs, Ss, z.curr, z.priors, demos, clear = T)
}
\arguments{
\item{x}{A data frame of input data.}

\item{Cs}{Character vector of variable names containing student challenge measures.}

\item{z.curr}{Character scalar variable name containing current-grade/year z-score scaled scores.}

\item{z.priors}{Character vector of variable names containing prior-grade/year z-score scaled scores.}

\item{demos}{Character vector of variable names containing demographic dummy variables or numeric demographic variables.}

\item{clear}{Logical scalar indicating whether to clear the console for user updates.}
}
\value{
The argument \code{x} with descriptive statistics calculated from \code{x$x} added to \code{x$y}.
}
\description{
Attach Input Descriptive Statistics to Output
}
