% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_mgp.R
\encoding{UTF-8}
\name{calc_mgp}
\alias{calc_mgp}
\title{Calculate Mean Student Growth Percentile (MGP)}
\usage{
calc_mgp(x)
}
\arguments{
\item{x}{A data frame containing entity.id, year, subject, and SGPs.}
}
\value{
A data frame.
}
\description{
Calculate Mean Student Growth Percentile (MGP)
}
