% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_pS.R
\encoding{UTF-8}
\name{calc_pS}
\alias{calc_pS}
\title{Calculate Proportion Successful by Subject, Year, and Cut Score}
\usage{
calc_pS(x, min.n = 30, lo = 0.05, hi = 0.95)
}
\arguments{
\item{x}{A data frame containing variables \code{entity.id}, \code{year}, \code{subject}, \code{S1}, \code{S2}, and \code{S3}.}

\item{min.n}{The minimum number of students necessary for conducting an SRT analysis for an entity group.}

\item{lo, hi}{The success rates at or below which or at or above which, respectively, SRT analysis is not useful.}
}
\value{
A data frame with the following variables:

\if{html}{\out{<div class="sourceCode">}}\preformatted{  entity.id  Entity ID number
  year       School year
  subject    Subject matter
  pS2        Proportion meeting cut score 2
  pS3        Proportion meeting cut score 3
  pS4        Proportion meeting cut score 4
  cond2      Condition code for cut score 2
  cond3      Condition code for cut score 3
  cond4      Condition code for cut score 4
}\if{html}{\out{</div>}}
}
\description{
Assumes three cut scores generating three percent successful measures.
}
