% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_Ss_from_ALs.R
\encoding{UTF-8}
\name{calc_Ss_from_ALs}
\alias{calc_Ss_from_ALs}
\title{Calculate Success Indicators from Achievement Levels}
\usage{
calc_Ss_from_ALs(x, ach.lev)
}
\arguments{
\item{x}{A data frame containing a current-year achievement level variable}

\item{ach.lev}{Character scalar name of the variable in \code{x} containing achievement levels.}
}
\value{
A data frame with as many new columns as there are achievement levels above \code{1}.
}
\description{
Calculate Success Indicators from Achievement Levels
}
