% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ss_to_z.R
\encoding{UTF-8}
\name{calc_ss_to_z}
\alias{calc_ss_to_z}
\title{Recalculate Scaled Scores as \eqn{z}-Scores for Use in Value-Added Modeling}
\usage{
calc_ss_to_z(x, curr.ss, priors, clear = T)
}
\arguments{
\item{x}{A data frame containing variables named \code{'year'}, \code{'subject'}, and \code{'grade'} as well as variables named in \code{curr.ss} and \code{priors}.}

\item{curr.ss}{Character scalar name of the variable containing current-year scaled scores.}

\item{priors}{Character vector of names of the variables containing prior scaled scores.}

\item{clear}{Logical scalar indicating whether to clear the console for user updates.}
}
\value{
The argument \code{x} with \eqn{z}-scores attached for the variables named in \code{ss} and \code{priors}.
}
\description{
Recalculate Scaled Scores as \eqn{z}-Scores for Use in Value-Added Modeling
}
