% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cors_adj_years.R
\encoding{UTF-8}
\name{cors_adj_years}
\alias{cors_adj_years}
\title{Calculate Adjacent-Year Correlations of \eqn{E}}
\usage{
cors_adj_years(x, lo = 0.1, hi = 0.9, min.n = 20)
}
\arguments{
\item{x}{The output of a call to \code{\link{srt}}.}
}
\value{
A data frame
}
\description{
Calculate Adjacent-Year Correlations of \eqn{E}
}
