% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cors_adj_years.R
\encoding{UTF-8}
\name{cors_vam}
\alias{cors_vam}
\title{Calculate Correlations of \eqn{E} with VAM}
\usage{
cors_vam(x, lo = 0.1, hi = 0.9, min.n = 20)
}
\arguments{
\item{x}{The output of a call to \code{\link{srt}}.}

\item{vam}{The output of a call to \code{\link{hlm_vam}}.}

\item{mgp}{The output of a call to \code{\link{calc_mgp}}.}
}
\value{
A data frame.
}
\description{
Calculate Correlations of \eqn{E} with VAM
}
