% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_srt.R
\encoding{UTF-8}
\name{est_srt}
\alias{est_srt}
\title{Estimate a Student Response Theory Model for a Single Educational Entity-Group}
\usage{
est_srt(
  x,
  C,
  S,
  lb,
  ub,
  max.jack = 1e+05,
  min.n = 30,
  clear = T,
  head = NULL,
  dir = getwd(),
  save.n = 500,
  say.n = 50,
  no.pickup = F,
  n.core = 6,
  lo = NULL,
  hi = NULL,
  E.known = NULL
)
}
\arguments{
\item{x}{A data frame for educational entity \eqn{k}.}

\item{C}{Character scalar name of the variable in \code{x} containing values of \eqn{C_{jk}}, or the challenge posed by students in eliciting student success.}

\item{S}{Character scalar name of the variable in \code{x} containing values of \eqn{S_{jk}}, or the 0/1 student success indicator.}

\item{max.jack}{The maximum number of jackknifed estimations to conduct per entity in calculating point estimates and standard errors of \eqn{E_k} and \eqn{D_k}.}

\item{min.n}{Minimum number of students required for an entity to get an estimate.}

\item{clear}{Logical scalar indicating whether to clear the console for each round of estimation.}

\item{head}{Optional character scalar user update header.}

\item{dir}{Character scalar directory in which to save interim results.}

\item{save.n}{Integer scalar number of entity-groups after which to save interim results.}

\item{say.n}{Integer scalar number of entity-groups after which to provide a user update.}

\item{entity.id}{Character scalar name of the variable in \code{x} containing entity IDs.}

\item{year}{Character scalar name of the variable in \code{x} containing year identifiers.}

\item{subject}{Character scalar name of the variable in \code{x} containing subject identifiers.}
}
\value{
A list of the following structure:

\if{html}{\out{<div class="sourceCode">}}\preformatted{  $ x            An extract of the data frame in argument [x] containing complete
                   cases of only variables entity.id, year, subject, [S], and [C].
  $ y            A data frame with the following variables:
    $ entity.id    Entity-group ID variable 1
    $ year         Entity-group ID variable 2
    $ subject      Entity-group ID variable N
    $ ns           The number of students with complete data in the entity
    $ nj           The number of jackknifed samples drawn for the entity
    $ E            Bounded, jackknifed estimate of entity capacity
    $ D            Bounded, jackknifed estimate of entity discrimination
    $ se.E         Bounded, jackknifed standard error of [E]
    $ se.D         Bounded, jackknifed standard error of [D]
    $ e            Single-sample estimate of entity capacity.
    $ d            Single-sample estimate of entity discrimination
    $ u            Unbounded, jackknifed estimate of entity capacity
    $ se.u         Unbounded, jackknifed standard error of [ue]
    $ neg.D        TRUE/FALSE indicator that D < 0
    $ neg.d        TRUE/FALSE indicator that d < 0
  $ args         A list containing preferences/settings of the following structure:
    $ S            Character scalar argument [S]
    $ C            Character scalar argument [C]
    $ max.jack     Integer scalar argument [max.jack]
    $ min.n        Integer scalar argument [min.n]
}\if{html}{\out{</div>}}
}
\description{
An entity-group is the collection of students with each unique combination of identifying variable values. For example, if estimates are desired for each subject, grade, and year for each school building, then there are four variables whose unique combinations identify an entity-group: subject, grade, year, and school building ID.
\cr\cr
This function estimates, for each entity-group, the following SRT model:
\deqn{P\left(S_{jk} = 1 \mid C_{jk}, E_k, D_k \right) = 1 / \big(1 + e^{-D_k\left(E_k - C_{jk} \right)}\big)}
where
\tabular{lrll}{
   \tab \eqn{S_{jk}} \tab    \tab is a 0/1 indicator of student success for student \eqn{j} in entity \eqn{k},                                           \cr
   \tab \eqn{C_{jk}} \tab    \tab is the numeric degree of challenge posed in eliciting student success for student \eqn{j} in entity \eqn{k},           \cr
   \tab \eqn{E_k}    \tab    \tab is entity \eqn{k}'s capacity to elicit student success on the same scale as \eqn{C_i},                                 \cr
   \tab \eqn{D_k}    \tab    \tab is entity \eqn{k}'s discrimination in not eliciting vs. eliciting student success below and above \eqn{E_j}, respectively.
}
To ensure that a model can be estimated for each entity, for each model estimated, two student records are imputed:
\itemize{
\item A student \eqn{l} for whom \eqn{S_{lk} = 1} and \eqn{C_{lk} = LB_C + R \sim U(0, min_{C_k} - LB_C)}.
\item A student \eqn{m} for whom \eqn{S_{mk} = 0} and \eqn{C_{mk} = UB_C - R \sim U(0, UB_C - max_{C_k})}.
}
where \eqn{R \sim U(A, B)} is a single uniform random value sampled from the interval \eqn{\left[A, B\right]}.
}
