% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlm_vam.R
\encoding{UTF-8}
\name{hlm_vam}
\alias{hlm_vam}
\title{Conduct Hierarchical Linear Value Added Model Analyses}
\usage{
hlm_vam(x, z.curr, z.priors, demos = NULL, clear = T)
}
\arguments{
\item{x}{A data frame containing entity.id, year, subject, and z-scored scaled scores.}

\item{z.curr}{Character scalar name of the variable containing \eqn{z}-scored current scaled scores.}

\item{z.priors}{Character vector of names of variables containing \eqn{z}-scored prior scaled scores.}

\item{demos}{Optional character vector of names of variables containing demographics.}

\item{clear}{Logical scalar indicating whether to clear the console for user updates.}
}
\value{
A data frame.
}
\description{
Conducts one analysis per unique combination of year and subject.
}
