% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest_data.R
\encoding{UTF-8}
\name{ingest_cuts}
\alias{ingest_cuts}
\title{Ingest a Data Frame of Threshold/Cut Scores}
\usage{
ingest_cuts(x = NULL, d = NULL, file = NULL)
}
\value{
A data frame.
}
\description{
Ingest a Data Frame of Threshold/Cut Scores
}
\details{
The data frame ingested either as the argument \code{x} or as read from disk must have variables associated with year, subject, grade, achievement level, and threshold score.
}
