% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest_data.R
\encoding{UTF-8}
\name{ingest_student_data}
\alias{ingest_student_data}
\title{Ingest a Student-Level Data Frame to Be Used in SRT Analysis}
\usage{
ingest_student_data(x = NULL, d = NULL, file = NULL)
}
\arguments{
\item{x}{Optional \code{\link[uj:atm_dtf]{atomic data frame}} to be ingested. When \code{x} is \code{NULL}, a delimited data file is read from disk.}

\item{d}{Optional single-character scalar text delimiter. When both \code{x} and \code{d} are \code{NULL}, the user is asked to select from among a set of delimiters.}

\item{file}{Optional character scalar path to a file. When both \code{x} and \code{file} are \code{NULL}, the user is asked to select a data file using the operating system's file selection dialog box.}
}
\value{
A data frame.
}
\description{
Ingest a Student-Level Data Frame to Be Used in SRT Analysis
}
\details{
The data frame ingested either as the argument \code{x} or as read from disk must have variables associated with entity ID, year, subject, and grade.
}
