% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_caterpillars.R
\encoding{UTF-8}
\name{plot_caterpillars}
\alias{plot_caterpillars}
\title{Build a Caterpillar Plot for Entity Group Capacity}
\usage{
plot_caterpillars(
  x,
  n.se = 1,
  lo = 0.05,
  hi = 0.95,
  min.n = 20,
  dir = getwd(),
  sub.dir = "plots",
  sub.sub = "caterpillars",
  w = 6.5,
  h = 8.5,
  dpi = 500,
  dev = "png",
  clear = T
)
}
\arguments{
\item{x}{The output of a call to \code{\link{srt}}.}

\item{n.se}{Numeric scalar number of standard errors above and below point estimates to go in forming confidence intervals.}

\item{min.n}{Integer scalar minimum number of students for an entity-group to be plotted.}

\item{dir}{Character scalar directory in which to save plots.}

\item{sub.dir}{Character scalar sub-directory in which to save plots.}

\item{sub.sub}{Character scalar sub-sub-directory in which to save plots.}

\item{w, h}{Numeric scalar width and height of plots in inches.}

\item{dpi}{Integer scalar dots per inch of resolution.}

\item{dev}{Character scalar graphics device to use in saving plots.}

\item{clear}{Logical scalar indicating whether to clear the console for user updates.}

\item{Cs}{Character vector of names of the student challenge variables for which to plot entity-group capacity.}
}
\value{
\code{NULL}. Called for the side effect of building and saving plots.
}
\description{
Build a Caterpillar Plot for Entity Group Capacity
}
