% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_e_vs_E.R
\encoding{UTF-8}
\name{plot_e_vs_E}
\alias{plot_e_vs_E}
\title{Plot Single-Sample vs. Jackknifed Estimates of \eqn{E}}
\usage{
plot_e_vs_E(
  x,
  dir = getwd(),
  sub.dir = "plots",
  sub.sub = "e_vs_E",
  w = 6.5,
  h = 4,
  dpi = 500,
  dev = "png",
  clear = T
)
}
\arguments{
\item{x}{The input data for a call to \code{\link{srt}}.}

\item{dir}{Character scalar directory in which to save plots.}

\item{sub.dir}{Character scalar sub-directory in which to save plots.}

\item{sub.sub}{Character scalar sub-sub-directory in which to save plots.}

\item{w, h}{Numeric scalar width and height of plots in inches.}

\item{dpi}{Integer scalar dots per inch of resolution.}

\item{dev}{Character scalar graphics device to use in saving plots.}

\item{clear}{Logical scalar indicating whether to clear the console for user updates.}
}
\value{
\code{NULL}. Called for the side effect of building and saving plots.
}
\description{
Plot Single-Sample vs. Jackknifed Estimates of \eqn{E}
}
