% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_entity_group.R
\encoding{UTF-8}
\name{plot_entity_group}
\alias{plot_entity_group}
\title{Build and Save a Student Success Plots for a Single Entity Group}
\usage{
plot_entity_group(
  x,
  Ss,
  Cs,
  entity.id = NULL,
  min.n = 30,
  lo = 0.05,
  hi = 0.95,
  dir = getwd(),
  sub.dir = "plots",
  sub.sub = "entities",
  w = 6.5,
  h = 4.5,
  dpi = 500,
  dev = "png",
  clear = T,
  mask = T
)
}
\arguments{
\item{x}{The output of a call to \code{\link{srt}}.}

\item{Ss}{Character vector of student success variable names.}

\item{Cs}{Character vector of student challenge variable names of the same length as \code{S}.}

\item{entity.id}{Optional character scalar entity ID.}

\item{min.n}{Integer scalar minimum number of students for an entity-group to be plotted.}

\item{dir}{Character scalar directory in which to save plots.}

\item{sub.dir}{Character scalar sub-directory in which to save plots.}

\item{sub.sub}{Character scalar sub-sub-directory in which to save plots.}

\item{w, h}{Numeric scalar width and height of plots in inches.}

\item{dpi}{Integer scalar dots per inch of resolution.}

\item{dev}{Character scalar graphics device to use in saving plots.}

\item{clear}{Logical scalar indicating whether to clear the console for user updates.}
}
\value{
\code{NULL}. Called for the side effect of building and saving plots
}
\description{
Build and Save a Student Success Plots for a Single Entity Group
}
