% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_srt.R
\encoding{UTF-8}
\name{plot_sim_srts_p}
\alias{plot_sim_srts_p}
\title{Plot Simulated SRT Recovery Results}
\usage{
plot_sim_srts_p(
  x,
  lo = 0.05,
  hi = 0.95,
  min.n = 30,
  dir = getwd(),
  sub.dir = "plots",
  sub.sub = "simulation",
  w = 6.5,
  h = 8.2,
  dpi = 500,
  dev = "png",
  clear = T
)
}
\arguments{
\item{x}{The output of a call to \code{\link{sim_srts_prob}}.}

\item{dir}{Character scalar directory in which to save plots.}

\item{sub.dir}{Character scalar sub-directory in which to save plots.}

\item{sub.sub}{Character scalar sub-sub-directory in which to save plots.}

\item{w, h}{Numeric scalar width and height of plots in inches.}

\item{dpi}{Integer scalar dots per inch of resolution.}

\item{dev}{Character scalar graphics device to use in saving plots.}

\item{clear}{Logical scalar indicating whether to clear the console for user updates.}
}
\value{
\code{NULL}. Called for the side effect of building and saving plots.
}
\description{
Plot Simulated SRT Recovery Results
}
