% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/review_EvE.R
\encoding{UTF-8}
\name{review_EvE}
\alias{review_EvE}
\title{Graphically Review Different Variants of Entity-Group Capacity}
\usage{
review_EvE(x, ID, C)
}
\arguments{
\item{ID}{Character vector of entity-group identifying variable names.}

\item{C}{Character vector of student challenge variable names.}
}
\value{
\code{NULL}. Called for the side effect of plotting two-variable relationships for review.
}
\description{
Graphically Review Different Variants of Entity-Group Capacity
}
