% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_srt.R
\encoding{UTF-8}
\name{sim_srts_p}
\alias{sim_srts_p}
\title{Simulate an SRT Analysis to Evaluate the Effect of Partial Jackknifing on Standard Errors of Probability-Based Entity Capacity Measures}
\usage{
sim_srts_p(
  ni = 30,
  mn.mn.z = 0,
  sd.mn.z = 0.6,
  mn.sd.z = 0.85,
  sd.sd.z = 0.15,
  mn.E = 50,
  sd.E = 15,
  lo = NULL,
  hi = NULL,
  z.cuts = round(seq(-2, 2, by = 0.1), 1),
  r.prev.curr = 0.8,
  save.n = 500,
  say.n = 50,
  clear = T,
  dir = getwd(),
  n.core = 6,
  max.jack = 5000
)
}
\arguments{
\item{ni}{Positive integer scalar number of iterations per sample size.}

\item{mn.mn.z}{Numeric scalar mean of entity-aggregate mean current-year z-scores.}

\item{sd.mn.z}{Numeric scalar standard deviation of entity-aggregate mean current-year z-scores.}

\item{mn.sd.z}{Numeric scalar mean of entity-aggregate standard deviations of current-year z-scores.}

\item{sd.sd.z}{Numeric scalar standard deviation of entity-aggregate standard deviations of current-year z-scores.}

\item{mn.E}{Numeric scalar mean entity capacity.}

\item{sd.E}{Numeric scalar standard deviation of entity capacity.}

\item{lo, hi}{Numeric scalar lower and upper bound on proportion student success to identify for which entities to estimate capacity.}

\item{r.prev.curr}{Numeric scalar correlation between current-year and previous-year scores.}

\item{clear}{Logical scalar indicating whether to clear the console for major user updates.}

\item{dir}{Character scalar directory in which to save interim results.}

\item{n.core}{Integer scalar number of cores to use for parallel processing.}

\item{maj.jacks}{Integer vector of maximum jackknife draws for which to run analyses.}
}
\value{
A data frame with the following structure:

\if{html}{\out{<div class="sourceCode">}}\preformatted{  $ entity.id  Entity's integer ID number
  $ ns         Sample size (number of students)
  $ E.known    Entity's known capacity value
  $ D.known    Entity's known discrimination value
  $ nj.all     Integer number of jackknife draws for all possible pairs
  $ nj.max     Integer number of jackknife draws for bounded number of pairs
  $ E.hat.all  Estimated entity capacity using all possible jackknife draws
  $ E.hat.max  Estimated entity capacity using a bounded number of jackknife draws
  $ D.hat.all  Estimated entity discrimination using all possible jackknife draws
  $ D.hat.max  Estimated entity discrimination using a bounded number of jackknife draws
  $ se.all     Estimated SE(E) using all possible jackknife draws
  $ se.max     Estimated SE(E) using a bounded number of jackknife draws
  $ d          Proportional deflation of SE(E) associated with using a bounded number of jackknife draws
}\if{html}{\out{</div>}}
}
\description{
Simulate an SRT Analysis to Evaluate the Effect of Partial Jackknifing on Standard Errors of Probability-Based Entity Capacity Measures
}
