% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_srt.R
\encoding{UTF-8}
\name{sim_srts_z}
\alias{sim_srts_z}
\title{Simulate an SRT Analysis}
\usage{
sim_srts_z(
  ne = 1000,
  mn.mn.z = 0,
  sd.mn.z = 0.6,
  mn.sd.z = 0.85,
  sd.sd.z = 0.15,
  mn.E = 0,
  sd.E = 1.5,
  lb = -6,
  ub = 6,
  r.prev.curr = 0.8,
  z.cuts = round(seq(-2, 2, by = 0.1), 1),
  clear = T,
  save.n = 100,
  say.n = 10,
  n.core = 6
)
}
\arguments{
\item{ne}{Integer scalar number of entities for which to simulate scaled scores}

\item{clear}{Logical scalar indicating whether to clear the console with each new set of student success and challenge variables.}

\item{save.n}{Integer scalar number of entity-groups after which to save interim results.}

\item{say.n}{Integer scalar number of entity-groups after which to provide a user update.}

\item{prob}{Logical scalar indicating whether to calculate student challenge (\eqn{C}) as probability of meeting the cut score (prob = TRUE) vs. as the \eqn{z}-score distance of previous scaled score to cut score.}

\item{cut}{Numeric scalar \eqn{z}-score cut score}

\item{cuts}{Numeric vector of \eqn{z}-score cut scores.}
}
\value{
A data.frame with the following variables:

\if{html}{\out{<div class="sourceCode">}}\preformatted{  $ year         Year with value [1]
  $ subject      Subject area with value ['x'].
  $ entity.id    Entity identifier
  $ C            Character scalar type of challenge from ["C"]
  $ S            Character scalar type of success from ["S"]
  $ ns           Number of students in the entity
  $ nj           Number of jackknife draws
  $ E            Jackknifed estimate of entity capacity (bounded)
  $ D            Jackknifed estimate of entity discrimination
  $ se.E         Jackknifed standard error of entity capacity (bounded)
  $ se.D         Jackknifed standard error of entity discrimination
  $ d            Single-sample estimate of entity discrimination
  $ e            Single-sample estimate of entity capacity (bounded)
  $ u            Jackknifed estimate of entity capacity (unbounded)
  $ se.u         Jackknifed estimated standard error of entity capacity (unbounded)
  $ neg.D        Logical indicator of whether [D < 0]
  $ neg.d        Logical indicator of whether [d < 0]
  $ vam          HLM value-added model estimate
  $ pS           Proportion successful (S = 1) in the entity
  $ mnC          Mean challenge (C) value in this entity
}\if{html}{\out{</div>}}

A data frame with one row per entity and the following variables

\if{html}{\out{<div class="sourceCode">}}\preformatted{  $ entity.id   Simulated entity ID
  $ E           Known entity capacity value
  $ z.cut       Z-scored cut score
  $ E.hat.p     Probability-based entity capacity estimate (with C based on probability of missing the cut score)
  $ year        Variable indicating year (all values equal to [1])
  $ subject     Variable indicating subject (all values equal to ['x'])
  $ pS          Proportion of students meeting the cut score
}\if{html}{\out{</div>}}
}
\description{
Simulate an SRT Analysis

Simulate SRT Analyses for Each of Sequence of Cut Scores
}
