% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{srt_example}
\alias{srt_example}
\title{Get path to srt example}
\usage{
srt_example()
}
\value{
The path or name to a example \code{.srt} file.
}
\description{
srt comes bundled with a number of sample files in its \code{inst/extdata}
directory. This function make them easy to access.
}
\details{
\emph{It's a Wonderful Life} (1946) entered the public domain in 1974.
}
\examples{
srt_example()
}
