% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{srt_seconds}
\alias{srt_seconds}
\alias{srt_index}
\alias{srt_text}
\title{Parse components of a subtitle file}
\usage{
srt_seconds(x)

srt_index(x)

srt_text(x, collapse = "\\n")
}
\arguments{
\item{x}{A character vector with the lines of an \code{.srt} file.}

\item{collapse}{The character with which to separate subtitle lines.}
}
\description{
Parse components of a subtitle file
}
