% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{srt_example}
\alias{srt_example}
\title{Get path to srt example}
\usage{
srt_example(file = NULL)
}
\arguments{
\item{file}{Name or number of file. If \code{NULL}, file names are be listed.}
}
\value{
The path or name to example \code{.srt} files.
}
\description{
srt comes bundled with a number of sample files in its \code{inst/extdata}
directory. This function make them easy to access.
}
\examples{
srt_example()
srt_example(1)
srt_example("toy-story.en.srt")
}
